mlEngine/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
pandas
faiss-cpu
sentence-transformers
huggingface-hub